Move `DataComponentAdapter`
