+++ paper/src/main/java/com/denizenscript/denizen/paper/datacomponents/ComponentAdaptersRegistry.java
@@ -1,7 +1,5 @@
 package com.denizenscript.denizen.paper.datacomponents;
 
-import com.denizenscript.denizen.paper.utilities.DataComponentAdapter;
-
 public class ComponentAdaptersRegistry {
 
     public static void register() {
